fixed some little float issues, etc

+++ Inc/oled.h
@@ -22,27 +22,27 @@ void OLED_init(I2C_HandleTypeDef* hi2c);
 
 void initUIFunctionPointers(void);
 
-void setLED_Edit(uint8_t onOff);
+void setLED_Edit(int onOff);
 
-void setLED_USB(uint8_t onOff);
+void setLED_USB(int onOff);
 
-void setLED_1(uint8_t onOff);
+void setLED_1(int onOff);
 
-void setLED_2(uint8_t onOff);
+void setLED_2(int onOff);
 
-void setLED_A(uint8_t onOff);
+void setLED_A(int onOff);
 
-void setLED_B(uint8_t onOff);
+void setLED_B(int onOff);
 
-void setLED_C(uint8_t onOff);
+void setLED_C(int onOff);
 
-void setLED_leftout_clip(uint8_t onOff);
+void setLED_leftout_clip(int onOff);
 
-void setLED_rightout_clip(uint8_t onOff);
+void setLED_rightout_clip(int onOff);
 
-void setLED_leftin_clip(uint8_t onOff);
+void setLED_leftin_clip(int onOff);
 
-void setLED_rightin_clip(uint8_t onOff);
+void setLED_rightin_clip(int onOff);
 
 int getCursorX(void);
 
@@ -52,9 +52,9 @@ void OLED_writePreset(void);
 
 void OLED_writeEditScreen(void);
 
-void OLED_writeKnobParameter(uint8_t whichParam);
+void OLED_writeKnobParameter(int whichParam);
 
-void OLED_writeButtonAction(uint8_t whichButton, uint8_t whichAction);
+void OLED_writeButtonAction(int whichButton, int whichAction);
 
 void OLED_writeTuning(void);
 
@@ -64,27 +64,27 @@ void OLEDclear(void);
 
 void OLEDclearLine(OLEDLine line);
 
-void OLEDwriteString(const char* myCharArray, uint8_t arrayLength, uint8_t startCursor, OLEDLine line);
+void OLEDwriteString(const char* myCharArray, int arrayLength, int startCursor, OLEDLine line);
 
-void OLEDwriteLine(const char* myCharArray, uint8_t arrayLength, OLEDLine line);
+void OLEDwriteLine(const char* myCharArray, int arrayLength, OLEDLine line);
 
-void OLEDwriteInt(uint32_t myNumber, uint8_t numDigits, uint8_t startCursor, OLEDLine line);
+void OLEDwriteInt(uint32_t myNumber, int numDigits, int startCursor, OLEDLine line);
 
-void OLEDwriteIntLine(uint32_t myNumber, uint8_t numDigits, OLEDLine line);
+void OLEDwriteIntLine(uint32_t myNumber, int numDigits, OLEDLine line);
 
-void OLEDwritePitch(float midi, uint8_t startCursor, OLEDLine line, uint8_t showCents);
+void OLEDwritePitch(float midi, int startCursor, OLEDLine line, int showCents);
 
-void OLEDwritePitchClass(float midi, uint8_t startCursor, OLEDLine line);
+void OLEDwritePitchClass(float midi, int startCursor, OLEDLine line);
 
-void OLEDwritePitchLine(float midi, OLEDLine line, uint8_t showCents);
+void OLEDwritePitchLine(float midi, OLEDLine line, int showCents);
 
-void OLEDwriteFixedFloat(float input, uint8_t numDigits, uint8_t numDecimal, uint8_t startCursor, OLEDLine line);
+void OLEDwriteFixedFloat(float input, int numDigits, int numDecimal, int startCursor, OLEDLine line);
 
-void OLEDwriteFixedFloatLine(float input, uint8_t numDigits, uint8_t numDecimal, OLEDLine line);
+void OLEDwriteFixedFloatLine(float input, int numDigits, int numDecimal, OLEDLine line);
 
-void OLEDwriteFloat(float input, uint8_t startCursor, OLEDLine line);
+void OLEDwriteFloat(float input, int startCursor, OLEDLine line);
 
-void OLEDdrawFloatArray(float* input, float min, float max, uint8_t size, uint8_t offset, uint8_t startCursor, OLEDLine line);
+void OLEDdrawFloatArray(float* input, float min, float max, int size, int offset, int startCursor, OLEDLine line);
 
-int16_t OLEDgetCursor(void);
+int OLEDgetCursor(void);
 #endif /* OLED_H_ */
